requirements.txt
authy==2.2.6
better_profanity==0.7.0
Django
djangorestframework==3.14.0
djangorestframework_simplejwt==5.3.0
notification==0.2.1
django-cors-headers==3.13.0
pusher==3.3.2
gunicorn
whitenoise
pillow
psycopg2
django-storages 
boto3
setuptools

